fixed: start wasm loop after loading vgg file
