+++ src/Entry/WASM/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(vgg_runtime "main.cpp" "WasmBindings.cpp")
+add_executable(vgg_runtime "main.cpp")
   set_target_properties(vgg_runtime PROPERTIES
     COMPILE_FLAGS "-s USE_SDL=2 -fexceptions"
     LINK_FLAGS "\
